summaries: reorganize
